+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## [0.4.4] - Unreleased
 ### Added
-* WorkspaceFile attribute to Pipeline class to customize acces controls
+* WorkspaceFile attribute to Pipeline class to customize access controls
 ### Fixed
 * n/a
 ### Updated
 * Changed pipeline JSON file permission from `account users` to `users`
+* Smart join to support coalesce of columns outside of the join
 ### Breaking changes
 * n/a