+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ### Added
 * WorkspaceFile attribute to Pipeline class to customize access controls
 ### Fixed
-* n/a
+* Spark dependencies
 ### Updated
 * Changed pipeline JSON file permission from `account users` to `users`
 * Smart join to support coalesce of columns outside of the join